Introduce first tests in the codebase
This commit introduces the first tests to the codebase with `jest`. They are
simple, but the primary objective here is adding `jest` so other people can
focus on writing tests. The new test exercises `censorText` (renamed to
`findAndReplaceText`), which is the simplest function and should serve as a
template for future tests.

Having tests means we can rapidly iterate over changes, without fear of breaking
other parts of the code. The alternative is super slow, specially with Obsidian
plugins: (1) change an algorithm, (2) save the file, (3) reload Obsidian, (4)
click a few button with a mouse, (5) check if result worked. This problems
repeats over and over. With the new tests, all we need to do instead is run

    npm run test

Obsidian makes testing hard because we don't have have modules available
outside of the app, such as `obsidian`. In case the codebase is coupled to
Obsidian, we won't be able to test anything because file won't be compile if it
has `import something from "obsidian"`. In order to be able to test the
`censorText` function, I was forced to extract it into another file so we avoid
the coupling with `obsidian`.

The ideal situation would be to have all code self-contained with our own
interfaces, without dealing with `obsidian` directly. Then some module wraps
around `obsidian`, ensuring 99% of the code is testable. The current scenario,
with `import ... obsidian` distributed in many files make it hard to test them.
Baby steps, I guess.

The new tests assert the basic functionality, such as regex patterns executing
ocurring in sequence and based off of `settings`. I also found some difficulty
setting my own regex in the settings, so I created an additional test which
serves as documentation for anyone who wants to use Jekyll in the future.

The `package.json` changes were minimal, but because `jest` requires `babel`,
adding `babel` made tons of changes in `package-lock.json`.