feat(embed): introduce "bake embed"
thanks to @mgmeyers

it allows to add the contents of the embed in the file directly, like in easy-bake plugin.

The contents is added "as in", without specific markdown (like >).

A new option is added in the settings to allow that.
To use it with frontmatter, user need to use "bake" or "include", like:

```yaml
removeEmbed: "bake"
or
embed:
 remove: "bake"
```

See #190