fix(autoclean): fix the autoclean **command** not disabled if the option is disabled
An autoclean condition was included in the main.ts to allow purgeNotesRemoteCallback when 'autoclean.enable' is true. Also, code in 'callback.ts' was adjusted for efficiency and log messaging was included to indicate when purge command is enabled. In 'settings.ts', command cleanup was set to run when autoclean is toggled. The 'utils/index.ts' was also updated for more efficient searching of 'shortRepo'.