fix(property): allow list with file link to be used as property value
Allow to send multiple file using the frontmatter or dataview syntax.
The plugin will automatically remove the `[[` and `]]` from the path and check the correspondate file.

Now you can use the following syntax:

```yaml
field:
  - "[[file1]]"
  - "[[file2]]"
```
To send the file `file1` and `file2` to the repository.
It will normally adjusted with Dataview