fix(cleanup): better handling image deletion
Now, backlink of image are fetched alongside the image when checking the files in Obsidian. This allow creating additional path for attachment when the backlinks have `attachment.folder` key in frontmatter. 

This will prevent deleting an image that should exist.

The `otherPaths` is a simple `string[]` with additional possible path. 

> [!warning]
> backlinks of an image can be considered as "long", and in larger vault it could create some long cleanup/lag.

Moreover, the Obsidian API was improved using `obsidian-typings`, improving and removing the usage of `//@ts-ignore` on some function.